admin/requirements.txt
flask>=2.3.0
flask-login>=0.6.0
flask-wtf>=1.2.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
gunicorn>=21.0.0
